abstract level factory cleaning

+++ GeometryGame/core/src/com/example/learning/game/gamelogic/components/BulletComponent.java
@@ -17,7 +17,7 @@ public class BulletComponent implements Component, Poolable {
     public long creationTime;
 
     /** Time life of bullet in milliseconds */
-    public float lifeTime;
+    public long lifeTime;
 
     /** list of broken line points on it's way */
     public java.util.List<Vector2> path = new ArrayList<>();
